requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
requests==2.32.5
beautifulsoup4==4.14.3
python-dotenv==1.2.1
schedule==1.2.2
psycopg[binary]==3.2.4